requirements.txt
asgiref==3.8.1
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.2.0
colorama==0.4.6
cryptography==43.0.1
dj-database-url==2.3.0
Django==5.1.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.16.0
msgpack==1.1.0
package_name==0.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
redis==6.1.0
sqlparse==0.4.4
typing_extensions==4.13.2
tzdata==2024.1
uvicorn==0.34.2
whitenoise==6.7.0
